// src/app/shared/models/userresponse.ts
export interface IUserresponse {

    userId: number;
    userName: string;
    roleID: string;
    icnId: number;
    rolename: string;
    loginId: number;
    UserName: string;
    Password: string;
    Pin: number;
    Question: string;
    Answer: string;
    FirstName: string;
    LastName: string;
    Email: string;
    privileges: any;
    LocationName: string;
    LocationCode: string;
}